server_visual/requirements.txt
pyserial
Flask-SQLAlchemy